Save work

+++ storage/kv/bbolt.go
@@ -10,7 +10,8 @@ import (
 var _ Store = (*BBoltStore)(nil)
 
 type BBoltStore struct {
-	db *bolt.DB
+	db        *bolt.DB
+	namespace [][]byte
 }
 
 func (store *BBoltStore) Begin(writable bool) (Transaction, error) {
@@ -20,52 +21,68 @@ func (store *BBoltStore) Begin(writable bool) (Transaction, error) {
 		return nil, fmt.Errorf("Could not begin transaction: %s", err.Error())
 	}
 
-	return &BBoltTransaction{transaction: transaction}, nil
+	return &BBoltTransaction{transaction: transaction, namespace: store.namespace}, nil
 }
 
-func (store *BBoltStore) BeginBucket(name []byte, writable bool) (BucketTransaction, error) {
-	transaction, err := store.Begin(writable)
+func (store *BBoltStore) SubStore(bucket []byte) Store {
+	namespace := make([][]byte, len(store.namespace)+1)
 
-	if err != nil {
-		return nil, fmt.Errorf("Could not begin bucket transaction: %s", err.Error())
-	}
+	copy(namespace, store.namespace)
+	namespace[len(namespace)-1] = bucket
 
-	bucket := transaction.Bucket(name)
-
-	bucketTransaction := &BBoltBucketTransaction{}
-	bucketTransaction.transaction = transaction.(*BBoltTransaction)
-	bucketTransaction.BBoltBucket = *bucket.(*BBoltBucket)
-
-	return bucketTransaction, nil
-}
-
-var _ BucketTransaction = (*BBoltBucketTransaction)(nil)
-
-type BBoltBucketTransaction struct {
-	BBoltBucket
-	transaction *BBoltTransaction
-}
-
-func (bucketTransaction *BBoltBucketTransaction) Commit() error {
-	return bucketTransaction.Commit()
-}
-
-func (bucketTransaction *BBoltBucketTransaction) Rollback() error {
-	return bucketTransaction.Rollback()
+	return &BBoltStore{db: store.db, namespace: namespace}
 }
 
 var _ Transaction = (*BBoltTransaction)(nil)
 
 type BBoltTransaction struct {
 	transaction *bolt.Tx
+	namespace   [][]byte
+}
+
+func (transaction *BBoltTransaction) bucket() *bolt.Bucket {
+	var bucket *bolt.Bucket
+
+	for _, b := range transaction.namespace {
+		bucket = transaction.transaction.Bucket(b)
+	}
+
+	return bucket
 }
 
 func (transaction *BBoltTransaction) Bucket(name []byte) Bucket {
-	return &BBoltBucket{bucket: transaction.transaction.Bucket(name)}
+	nsBucket := transaction.bucket()
+
+	if nsBucket == nil {
+		return &BBoltBucket{bucket: transaction.transaction.Bucket(name)}
+	}
+
+	return &BBoltBucket{bucket: nsBucket.Bucket(name)}
+}
+
+func (transaction *BBoltTransaction) Namespace(bucket []byte) Transaction {
+	namespace := make([][]byte, len(transaction.namespace)+1)
+
+	copy(namespace, transaction.namespace)
+	namespace[len(namespace)-1] = bucket
+
+	return &BBoltTransaction{
+		transaction: transaction.transaction,
+		namespace:   namespace,
+	}
 }
 
 func (transaction *BBoltTransaction) CreateBucket(name []byte) (Bucket, error) {
-	bucket, err := transaction.transaction.CreateBucket(name)
+	var bucket *bolt.Bucket
+	var err error
+
+	nsBucket := transaction.bucket()
+
+	if nsBucket == nil {
+		bucket, err = transaction.transaction.CreateBucket(name)
+	} else {
+		bucket, err = nsBucket.CreateBucket(name)
+	}
 
 	if err != nil {
 		return nil, fmt.Errorf("Could not create bucket: %s", err.Error())
@@ -75,7 +92,16 @@ func (transaction *BBoltTransaction) CreateBucket(name []byte) (Bucket, error) {
 }
 
 func (transaction *BBoltTransaction) CreateBucketIfNotExists(name []byte) (Bucket, error) {
-	bucket, err := transaction.transaction.CreateBucketIfNotExists(name)
+	var bucket *bolt.Bucket
+	var err error
+
+	nsBucket := transaction.bucket()
+
+	if nsBucket == nil {
+		bucket, err = transaction.transaction.CreateBucketIfNotExists(name)
+	} else {
+		bucket, err = nsBucket.CreateBucketIfNotExists(name)
+	}
 
 	if err != nil {
 		return nil, fmt.Errorf("Could not create bucket: %s", err.Error())
@@ -85,21 +111,43 @@ func (transaction *BBoltTransaction) CreateBucketIfNotExists(name []byte) (Bucke
 }
 
 func (transaction *BBoltTransaction) Cursor() Cursor {
-	return &BBoltCursor{cursor: transaction.transaction.Cursor()}
+	nsBucket := transaction.bucket()
+
+	if nsBucket == nil {
+		return &BBoltCursor{cursor: transaction.transaction.Cursor()}
+	}
+
+	return &BBoltCursor{cursor: nsBucket.Cursor()}
 }
 
 func (transaction *BBoltTransaction) ForEach(fn func(name []byte, bucket Bucket) error) error {
+	nsBucket := transaction.bucket()
+
+	if nsBucket != nil {
+		return nil
+	}
+
 	return transaction.transaction.ForEach(func(name []byte, b *bolt.Bucket) error {
 		return fn(name, &BBoltBucket{bucket: b})
 	})
 }
 
 func (transaction *BBoltTransaction) DeleteBucket(name []byte) error {
-	return transaction.transaction.DeleteBucket(name)
+	nsBucket := transaction.bucket()
+
+	if nsBucket == nil {
+		return transaction.transaction.DeleteBucket(name)
+	}
+
+	return nsBucket.DeleteBucket(name)
 }
 
 func (transaction *BBoltTransaction) Commit() error {
 	return transaction.transaction.Commit()
+}
+
+func (transaction *BBoltTransaction) OnCommit(cb func()) {
+	transaction.transaction.OnCommit(cb)
 }
 
 func (transaction *BBoltTransaction) Rollback() error {
@@ -108,6 +156,14 @@ func (transaction *BBoltTransaction) Rollback() error {
 
 func (transaction *BBoltTransaction) Size() int64 {
 	return transaction.transaction.Size()
+}
+
+func (transaction *BBoltTransaction) Snapshot() (io.ReadCloser, error) {
+	return nil, nil
+}
+
+func (transaction *BBoltTransaction) ApplySnapshot(io.Reader) error {
+	return nil
 }
 
 var _ Bucket = (*BBoltBucket)(nil)
@@ -166,14 +222,6 @@ func (bucket *BBoltBucket) NextSequence() (uint64, error) {
 
 func (bucket *BBoltBucket) Put(key []byte, value []byte) error {
 	return bucket.bucket.Put(key, value)
-}
-
-func (bucket *BBoltBucket) Snapshot() (io.ReadCloser, error) {
-	return nil, nil
-}
-
-func (bucket *BBoltBucket) ApplySnapshot(io.Reader) error {
-	return nil
 }
 
 var _ Cursor = (*BBoltCursor)(nil)
